requirements.txt
vasppy
numpy
matplotlib
pytest-lazy-fixture
